Add static method for Glob class
A static method 'glob' was added to the Glob class to allow for simplified path pattern matching and to be suitable as a dropin replacement for https://packagist.org/packages/webmozart/glob. The README was updated to reflect this change. In addition, new unit tests were written to ensure the functionality of the static method both with absolute and relative paths.